add testing
- Add test
- Changes

+++ src/main/java/hu/webuni/pl/kamarasd/dto/AddressDto.java
@@ -20,11 +20,11 @@ public class AddressDto {
 	
 	@NotBlank(message = "Housno need to be fill!")
 	public String houseno;
-	public float gpslong;
-	public float gpslat;
+	public double gpslong;
+	public double gpslat;
 	
-	public AddressDto(Long id, String country, String city, String street, String postal, String houseno, float gpslong,
-			float gpslat) {
+	public AddressDto(Long id, String country, String city, String street, String postal, String houseno, double gpslong,
+			double gpslat) {
 		super();
 		this.id = id;
 		this.country = country;
@@ -84,19 +84,19 @@ public class AddressDto {
 		this.houseno = houseno;
 	}
 	
-	public float getGpslong() {
+	public double getGpslong() {
 		return gpslong;
 	}
 	
-	public void setGpslong(float gpslong) {
+	public void setGpslong(double gpslong) {
 		this.gpslong = gpslong;
 	}
 	
-	public float getGpslat() {
+	public double getGpslat() {
 		return gpslat;
 	}
 	
-	public void setGpslat(float gpslat) {
+	public void setGpslat(double gpslat) {
 		this.gpslat = gpslat;
 	}
